requirements.txt
Django==1.11.13
django-bootstrap4==0.0.6
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
pytz==2018.4
